+++ pddns/providers/__init__.py
@@ -1,4 +1,5 @@
 """Imports the provider modules"""
 from .cloudflare import Cloudflare
+from .hurricaneElectric import HurricaneElectric
 
-__all__ = ["Cloudflare"]
+__all__ = ["Cloudflare", "HurricaneElectric"]